make VIA class out of Keyboard class
