+++ src/ruby/sig/keyboard.rbs
@@ -24,15 +24,6 @@ class Keyboard
 
   STANDARD_SPLIT: Symbol
   RIGHT_SIDE_FLIPPED_SPLIT: Symbol
-  
-  #VIA
-  VIA_FILENAME: String
-  VIA_IDs : Array[Symbol]
-  VIA_PROTOCOL_VERSION : Integer
-  REPORT_RAW_MAX_LEN : Integer
-  KEYBOARD_VALUES : Array[Symbol]
-  ID_UNHANDLED : Integer
-  @via_layer_count : Integer
 
   @SHIFT_LETTER_OFFSET_A: Integer
   @SHIFT_LETTER_OFFSET_UNDS: Integer
@@ -57,11 +48,8 @@ class Keyboard
     pushed_at:         Integer,
     released_at:       Integer
   }]
-  @mode_keys_via: Hash[Symbol, [ Symbol | Array[Symbol] | Proc | nil, Symbol | Proc, Integer | nil, Integer| nil ] ]
-  @enable_via: bool
-  @keymap_saved: bool
   @map_size: Integer
-  
+
   @switches: Array[Array[Integer]]
   @layer_names: Array[Symbol]
   @layer: Symbol
@@ -88,7 +76,7 @@ class Keyboard
   attr_accessor uart_pin: Integer
   attr_accessor layer_changed_delay: Integer
 
-  def append: (RGB | RotaryEncoder feature) -> void
+  def append: (RGB | RotaryEncoder | VIA feature) -> void
   def start_features: -> void
   def mutual_uart_at_my_own_risk= : (bool) -> void
   def uart_partner_push8 : (Integer) -> void
@@ -96,33 +84,7 @@ class Keyboard
   def uart_anchor : (Integer) -> Integer
   def uart_anchor_init : (Integer) -> void
   def uart_partner_init : (Integer) -> void
-  def report_hid : (Integer, String) -> void
 
-  # VIA Ruby
-  def raw_hid_report_received?: -> bool
-  def set_raw_hid_report_read: -> void
-  def get_last_received_raw_hid_report: -> Array[Integer]
-  def report_raw_hid : (Array[Integer]) -> bool
-  def save_keymap_via: -> void
-  def eval_val : (String) -> untyped
-  def raw_hid_receive : (Array[Integer]) -> Array[Integer]
-  def raw_hid_receive_kb : (Array[Integer]) -> Array[Integer]
-  def via_get_layer_name : (Integer) -> Symbol
-  def dynamic_keymap_set_buffer : (Array[Integer]) -> void
-  def dynamic_keymap_get_buffer : (Array[Integer]) -> Array[Integer]
-  def dynamic_keymap_get_keycode : (Integer, Integer, Integer) -> Integer
-  def dynamic_keymap_set_keycode : (Integer, Integer, Integer, Integer) -> void
-  def via_task: -> void
-  def prk_keycode_into_keysymbol : (Integer | Symbol | Proc) -> Symbol
-  def via_keycode_into_prk_keycode : (Integer) -> ( [Integer, Integer] | Symbol | Integer )
-  def get_composite_key : (Symbol) -> Array[Symbol]
-  def get_via_composite_keycode : (Symbol) -> Integer
-  def prk_keycode_into_via_keycode : (Integer | Symbol | Proc) -> Integer
-  def get_modifier_name : (Integer) -> String
-  def load_mode_keys_via: -> void
-  def define_mode_key_via : (Symbol key_name, [Symbol | Array[Symbol] | Proc | nil, Symbol | Proc, Integer?, Integer?] param) -> void
-  def convert_to_uint8_array : (String) -> Array[Integer]
-  
   def initialize: -> void
   def layer: () -> Symbol
   def set_anchor: (Symbol val) -> void
@@ -157,4 +119,8 @@ class Keyboard
   def entire_cols_size: () -> Integer
   def skip_position?: (Integer row_index, Integer col_index) -> bool
   def resolve_key_alias: (Symbol) -> Symbol
+  def get_layer: (Symbol | nil, Integer) -> ( Array[Array[Integer | Symbol | Proc]] | nil )
+  def set_layer: (Symbol, Array[Array[Integer | Symbol | Proc]]) -> void
+  def delete_mode_keys: (Symbol) -> void
+  def keycode_to_keysym: (Integer) -> Symbol
 end
