+++ src/ruby/sig/via.rbs
@@ -1,0 +1,42 @@
+class VIA
+  VIA_FILENAME: String
+  VIA_IDs : Array[Symbol]
+  VIA_PROTOCOL_VERSION : Integer
+  REPORT_RAW_MAX_LEN : Integer
+  KEYBOARD_VALUES : Array[Symbol]
+  ID_UNHANDLED : Integer
+  @layer_count : Integer
+  @mode_keys: Hash[Symbol, [ Symbol | Array[Symbol] | Proc | nil, Symbol | Proc, Integer | nil, Integer| nil ] ]
+  @composite_keys: Array[Symbol]
+  @keymap_saved: bool
+  @map_size: Integer
+  @kbd: Keyboard
+  @cols_size: Integer
+  @keymaps: Array[Array[Array[Integer]]]
+
+  attr_accessor kbd: Keyboard
+
+  def report_raw_hid : (Array[Integer]) -> bool
+  def save_keymap: -> void
+  def eval_val : (String) -> untyped
+  def raw_hid_receive : (Array[Integer]) -> Array[Integer]
+  def raw_hid_receive_kb : (Array[Integer]) -> Array[Integer]
+  def via_get_layer_name : (Integer) -> Symbol
+  def dynamic_keymap_set_buffer : (Array[Integer]) -> void
+  def dynamic_keymap_get_buffer : (Array[Integer]) -> Array[Integer]
+  def dynamic_keymap_get_keycode : (Integer, Integer, Integer) -> Integer
+  def dynamic_keymap_set_keycode : (Integer, Integer, Integer, Integer) -> void
+  def task: -> void
+  def via_keycode_into_prk_keycode : (Integer) -> ( [Integer, Integer] | Symbol | Integer )
+  def via_keycode_into_keysymbol : (Integer) -> Symbol
+  def expand_composite_key : (Symbol) -> Array[Symbol]
+  def get_via_composite_keycode : (Symbol) -> Integer
+  def prk_keycode_into_via_keycode : (Integer | Symbol | Proc) -> Integer
+  def get_modifier_name : (Integer) -> String
+  def load_mode_keys: -> void
+  def define_mode_key : (Symbol key_name, [Symbol | Array[Symbol] | Proc | nil, Symbol | Proc, Integer?, Integer?] param) -> void
+  def convert_to_uint8_array : (String) -> Array[Integer]
+  def start!: -> void
+  def sync_keymap : (?bool)-> void
+  def init_keymap: -> void
+end
